requirements.txt
streamlit
google-genai
python-dotenv
pymongo 
streamlit-folium
streamlit-echarts
yfinance